requirements.txt
tqdm
pyyaml
strsimpy
numpy
spacy
networkx
sklearn
